+++ src/utils/utils.js
@@ -1,0 +1,9 @@
+// Преобразование длительности фильма
+export function calcMovieDuration(duration) {
+  if (duration > 59) {
+    const hours = (duration - duration % 60) / 60;
+    const minutes = duration % 60;
+    return `${hours}ч ${minutes > 0 ? minutes + 'м' : ''}`
+  }
+  return `${duration}м`;
+}
